+++ java/src/hype/core/util/HColors.java
@@ -74,4 +74,8 @@ public class HColors {
 	public static int getBlue(int clr) {
 		return clr & 255;
 	}
+	
+	public static boolean isTransparent(int clr) {
+		return (clr & 0xFF000000) == 0;
+	}
 }