+++ CHANGELOG.md
@@ -1,10 +1,15 @@
+### build_20130611.1 (Jun 11, 2013)
+- hitbox detection for HPath now changes depending on mode and presence of fill color
+
 ### build_20130611.0 (Jun 11, 2013)
 - the method `HMath.bezierParam()` (both of them) no longer removes values in the params array that are not within the range `[0,1]`
 - renamed HVertexNEW to HVertex
 - renamed and replaced HVertexNEW to HVertex
+- new static HColors method: `isTransparent(int)`
 
 ### build_20130610.0 (Jun 10, 2013)
 _note: the following changes are not yet put into the pde files due to their experimental nature_
+
 - implemented adjustments for some edge cases in `HVertexNEW.intersectTest()`
 - various debugs and refactors for `HVertexNEW` and `HPathNEW`
 - new HMath methods: